+++ packages/gov-ui-kit/src/modules/components/proposal/proposalActions/actions/proposalActionTokenMint/proposalActionTokenMint.test.tsx
@@ -13,6 +13,7 @@ describe('<ProposalActionTokenMint /> component', () => {
     const createTestComponent = (props?: Partial<IProposalActionTokenMintProps>) => {
         const completeProps: IProposalActionTokenMintProps = {
             action: generateProposalActionTokenMint(),
+            index: 0,
             ...props,
         };